oracle-monitor/requirements.txt
fastapi
uvicorn
requests
kafka-python
pydantic
python-dotenv
supabase
kubernetes
jsonschema>=4.21.0
google-generativeai>=0.4.0